Hippo4j core configuration center validation judgment reconstruction (#374)
